use correct sizes for `tierNum` and `numMints`

+++ contracts/GudSoulbound721.sol
@@ -13,10 +13,10 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
 
     error InvalidNumTiers();
     error InsufficientValue();
-    error PublicMintingDisabled(uint256 tier);
-    error ExceedsMaxOwnership(uint256 tier);
-    error ExceedsMaxMerkleMintUses(uint256 tier);
-    error ExceedsMaxSupply(uint256 tier);
+    error PublicMintingDisabled(uint8 tier);
+    error ExceedsMaxOwnership(uint8 tier);
+    error ExceedsMaxMerkleMintUses(uint8 tier);
+    error ExceedsMaxSupply(uint8 tier);
     error NotOwner();
     error IncorrectOwnerSignature();
     error IncorrectMerkleProof();
@@ -24,11 +24,11 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
     error NoSuchToken(uint256 tokenId);
 
     Tier[] private _tiers;
-    mapping(uint256 => uint256) _numMinted;
-    mapping(address /*owner*/ => mapping(uint256 /*tier*/ => uint256 /*numOwned*/)) private _numOwned;
+    mapping(uint8 /*tier*/ => uint248 /*numMinted*/) _numMinted;
+    mapping(address /*owner*/ => mapping(uint8 /*tier*/ => uint248 /*numOwned*/)) private _numOwned;
     bytes32 private _mintMerkleRoot;
     bytes32 private _numMerkleRoots;
-    mapping(address /*owner*/ => mapping(uint256 /*tier*/ => uint256 /*numOwned*/))[] private _merkleMintUses;
+    mapping(address /*owner*/ => mapping(uint8 /*tier*/ => uint248 /*numOwned*/))[] private _merkleMintUses;
 
     function initialize(string memory name, string memory symbol, Tier[] memory tiers) external initializer {
         OwnableUpgradeable.__Ownable_init();
@@ -36,11 +36,11 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
         setTiers(tiers);
     }
 
-    function mint(address to, uint256[] calldata numMints) external payable {
+    function mint(address to, uint248[] calldata numMints) external payable {
         uint256 totalPrice = 0;
         for (uint i = 0; i < numMints.length; ++i) {
             if (numMints[i] != 0 && _tiers[i].publicPrice == type(uint256).max) {
-                revert PublicMintingDisabled(i);
+                revert PublicMintingDisabled(uint8(i));
             }
             totalPrice += _tiers[i].publicPrice * numMints[i];
         }
@@ -48,7 +48,7 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
     }
 
     function mint(
-        uint256[] calldata numMints,
+        uint248[] calldata numMints,
         MerkleMint calldata merkleMint,
         bytes32[] calldata merkleProof
     ) external payable {
@@ -59,14 +59,14 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
         ) == false) {
             revert IncorrectMerkleProof();
         }
-        mapping(uint256 => uint256) storage merkleMintUses = _merkleMintUses[_merkleMintUses.length - 1][merkleMint.to];
+        mapping(uint8 => uint248) storage merkleMintUses = _merkleMintUses[_merkleMintUses.length - 1][merkleMint.to];
         uint256 totalPrice = 0;
 
         for (uint i = 0; i < numMints.length; ++i) {
-            if (merkleMintUses[i] + numMints[i] > merkleMint.tierMaxMints[i]) {
-                revert ExceedsMaxMerkleMintUses(i);
+            if (merkleMintUses[uint8(i)] + numMints[i] > merkleMint.tierMaxMints[i]) {
+                revert ExceedsMaxMerkleMintUses(uint8(i));
             }
-            merkleMintUses[i] += numMints[i];
+            merkleMintUses[uint8(i)] += numMints[i];
 
             totalPrice += merkleMint.tierPrices[i] * numMints[i];
         }
@@ -115,7 +115,7 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
         return _tiers;
     }
 
-    function numMinted(uint256 tier) external view returns (uint256) {
+    function numMinted(uint8 tier) external view returns (uint248) {
         return _numMinted[tier];
     }
 
@@ -144,22 +144,22 @@ contract GudSoulbound721 is IGudSoulbound721, Soulbound721Upgradable, OwnableUpg
             return interfaceId == type(IGudSoulbound721).interfaceId || super.supportsInterface(interfaceId);
     }
 
-    function _mint(address to, uint256[] calldata numMints, uint256 totalPrice) private {
+    function _mint(address to, uint248[] calldata numMints, uint256 totalPrice) private {
         for (uint tierNum = 0; tierNum < numMints.length; ++tierNum) {
             Tier storage tier = _tiers[tierNum];
 
-            if (_numOwned[to][tierNum] + numMints[tierNum] > tier.maxOwnable) {
-                revert ExceedsMaxOwnership(tierNum);
+            if (_numOwned[to][uint8(tierNum)] + numMints[tierNum] > tier.maxOwnable) {
+                revert ExceedsMaxOwnership(uint8(tierNum));
             }
-            if (_numMinted[tierNum] + numMints[tierNum] > tier.maxSupply) {
-                revert ExceedsMaxSupply(tierNum);
+            if (_numMinted[uint8(tierNum)] + numMints[tierNum] > tier.maxSupply) {
+                revert ExceedsMaxSupply(uint8(tierNum));
             }
 
-            uint256 tokenId = (tierNum << 248) + _numMinted[tierNum];
+            uint256 tokenId = (tierNum << 248) + _numMinted[uint8(tierNum)];
             for (uint j = 0; j < numMints[tierNum]; ++j) {
                 _safeMint(to, ++tokenId);
             }
-            _numMinted[tierNum] += numMints[tierNum];
+            _numMinted[uint8(tierNum)] += numMints[uint8(tierNum)];
         }
 
         if (totalPrice > msg.value) {
